Refactor AI chat runtime tests
