+++ apps/android/feature/ai/src/test/java/com/flashcardsopensourceapp/feature/ai/AiChatRuntimeBootstrapAndLiveStreamTest.kt
@@ -1,0 +1,193 @@
+package com.flashcardsopensourceapp.feature.ai
+
+import com.flashcardsopensourceapp.data.local.model.AiChatContentPart
+import com.flashcardsopensourceapp.data.local.model.AiChatLiveEvent
+import com.flashcardsopensourceapp.data.local.model.AiChatRunTerminalOutcome
+import kotlinx.coroutines.ExperimentalCoroutinesApi
+import kotlinx.coroutines.flow.MutableSharedFlow
+import kotlinx.coroutines.flow.emptyFlow
+import kotlinx.coroutines.test.advanceUntilIdle
+import kotlinx.coroutines.test.runTest
+import org.junit.Assert.assertEquals
+import org.junit.Assert.assertFalse
+import org.junit.Assert.assertNull
+import org.junit.Assert.assertTrue
+import org.junit.Test
+
+@OptIn(ExperimentalCoroutinesApi::class)
+class AiChatRuntimeBootstrapAndLiveStreamTest {
+    @Test
+    fun bootstrapWhileVisibleWithActiveRunStartsLiveCollection() = runTest {
+        val repository = FakeAiChatRepository()
+        val liveEvents = MutableSharedFlow<AiChatLiveEvent>()
+        repository.bootstrapResponses += makeBootstrapResponse(
+            sessionId = "session-1",
+            activeRun = makeActiveRun(runId = "run-1", cursor = "5")
+        )
+        repository.liveFlows["run-1"] = liveEvents
+        val runtime = makeRuntime(scope = this, repository = repository)
+
+        runtime.onScreenVisible()
+        runtime.updateAccessContext(makeAccessContext(workspaceId = defaultTestWorkspaceId))
+        advanceUntilIdle()
+
+        assertEquals(listOf("run-1"), repository.attachRunIds)
+        assertEquals("run-1", runtime.state.value.activeRun?.runId)
+        assertTrue(runtime.state.value.isLiveAttached)
+
+        runtime.onScreenHidden()
+        advanceUntilIdle()
+    }
+
+    @Test
+    fun assistantMessageDoneKeepsActiveRunUntilRunTerminalCompleted() = runTest {
+        val repository = FakeAiChatRepository()
+        val liveEvents = MutableSharedFlow<AiChatLiveEvent>()
+        repository.startRunResponse = makeAcceptedStartRunResponse(
+            sessionId = "session-1",
+            activeRun = makeActiveRun(runId = "run-1", cursor = "0"),
+            messages = listOf(
+                makeUserMessage(
+                    content = listOf(AiChatContentPart.Text(text = "Hello")),
+                    timestampMillis = 1L
+                ),
+                makeAssistantStatusMessage(timestampMillis = 2L)
+            ),
+            composerSuggestions = emptyList()
+        )
+        repository.liveFlows["run-1"] = liveEvents
+        val runtime = makeRuntime(scope = this, repository = repository)
+
+        runtime.onScreenVisible()
+        runtime.updateDraftMessage(draftMessage = "Hello")
+        runtime.sendMessage()
+        advanceUntilIdle()
+
+        liveEvents.emit(
+            AiChatLiveEvent.AssistantMessageDone(
+                metadata = makeMetadata(runId = "run-1", cursor = "1"),
+                itemId = "item-1",
+                content = listOf(AiChatContentPart.Text(text = "Done")),
+                isError = false,
+                isStopped = false
+            )
+        )
+        advanceUntilIdle()
+
+        assertEquals("run-1", runtime.state.value.activeRun?.runId)
+        val lastMessageAfterDone = runtime.state.value.persistedState.messages.last()
+        assertEquals("item-1", lastMessageAfterDone.itemId)
+        assertEquals(false, lastMessageAfterDone.isError)
+        assertEquals(false, lastMessageAfterDone.isStopped)
+        assertEquals(
+            "Done",
+            (lastMessageAfterDone.content.single() as AiChatContentPart.Text).text
+        )
+
+        liveEvents.emit(
+            AiChatLiveEvent.RunTerminal(
+                metadata = makeMetadata(runId = "run-1", cursor = "2"),
+                outcome = AiChatRunTerminalOutcome.COMPLETED,
+                message = null,
+                assistantItemId = null,
+                isError = null,
+                isStopped = null
+            )
+        )
+        advanceUntilIdle()
+
+        assertNull(runtime.state.value.activeRun)
+        assertFalse(runtime.state.value.isLiveAttached)
+
+        runtime.onScreenHidden()
+        advanceUntilIdle()
+    }
+
+    @Test
+    fun runTerminalResetRequiredForcesBootstrapReload() = runTest {
+        val repository = FakeAiChatRepository()
+        val liveEvents = MutableSharedFlow<AiChatLiveEvent>()
+        repository.bootstrapResponses += makeBootstrapResponse(
+            sessionId = "session-1",
+            activeRun = makeActiveRun(runId = "run-1", cursor = "5")
+        )
+        repository.bootstrapResponses += makeBootstrapResponse(
+            sessionId = "session-1",
+            activeRun = null
+        )
+        repository.liveFlows["run-1"] = liveEvents
+        val runtime = makeRuntime(scope = this, repository = repository)
+
+        runtime.onScreenVisible()
+        runtime.updateAccessContext(makeAccessContext(workspaceId = defaultTestWorkspaceId))
+        advanceUntilIdle()
+
+        liveEvents.emit(
+            AiChatLiveEvent.RunTerminal(
+                metadata = makeMetadata(runId = "run-1", cursor = "6"),
+                outcome = AiChatRunTerminalOutcome.RESET_REQUIRED,
+                message = "refresh",
+                assistantItemId = null,
+                isError = null,
+                isStopped = null
+            )
+        )
+        advanceUntilIdle()
+
+        assertEquals(2, repository.loadBootstrapCalls)
+        assertNull(runtime.state.value.activeRun)
+        assertEquals(AiConversationBootstrapState.READY, runtime.state.value.conversationBootstrapState)
+        assertFalse(runtime.state.value.isLiveAttached)
+    }
+
+    @Test
+    fun acceptedRunDoesNotAttachLiveWhenScreenIsHidden() = runTest {
+        val repository = FakeAiChatRepository()
+        repository.startRunResponse = makeAcceptedStartRunResponse(
+            sessionId = "session-1",
+            activeRun = makeActiveRun(runId = "run-1", cursor = "0"),
+            messages = listOf(
+                makeUserMessage(
+                    content = listOf(AiChatContentPart.Text(text = "Hello")),
+                    timestampMillis = 1L
+                ),
+                makeAssistantStatusMessage(timestampMillis = 2L)
+            ),
+            composerSuggestions = emptyList()
+        )
+        val runtime = makeRuntime(scope = this, repository = repository)
+
+        runtime.updateDraftMessage(draftMessage = "Hello")
+        runtime.sendMessage()
+        advanceUntilIdle()
+
+        assertEquals("run-1", runtime.state.value.activeRun?.runId)
+        assertFalse(runtime.state.value.isLiveAttached)
+        assertTrue(repository.attachRunIds.isEmpty())
+    }
+
+    @Test
+    fun unexpectedLiveDetachTriggersBootstrapRecoveryError() = runTest {
+        val repository = FakeAiChatRepository()
+        repository.bootstrapResponses += makeBootstrapResponse(
+            sessionId = "session-1",
+            activeRun = makeActiveRun(runId = "run-1", cursor = "5")
+        )
+        repository.bootstrapResponses += makeBootstrapResponse(
+            sessionId = "session-1",
+            activeRun = makeActiveRun(runId = "run-1", cursor = "6")
+        )
+        repository.liveFlows["run-1"] = emptyFlow()
+        val runtime = makeRuntime(scope = this, repository = repository)
+
+        runtime.onScreenVisible()
+        runtime.updateAccessContext(makeAccessContext(workspaceId = defaultTestWorkspaceId))
+        advanceUntilIdle()
+
+        val alert = runtime.state.value.activeAlert as AiAlertState.GeneralError
+        assertEquals("AI live stream ended before message completion.", alert.message)
+        assertNull(runtime.state.value.activeRun)
+        assertFalse(runtime.state.value.isLiveAttached)
+        assertEquals(2, repository.loadBootstrapCalls)
+    }
+}
